рефакторинг: выносит константы в отдельный файл
