+++ src/utils/constant.js
@@ -1,0 +1,34 @@
+// Элементы HTML
+const BODY_ELEMENT = document.querySelector(`body`);
+
+// Классы CSS
+const HIDDEN_CLASS = `visually-hidden`;
+const ACTIVE_CLASS_FILTER = `main-navigation__item--active`;
+const ACTIVE_CLASS_SORT = `sort__button--active`;
+
+// Ключи
+const ENTER_KEY = `Enter`;
+const ESC_KEY = `Escape`;
+const COMMENT_FORM_FIELDS = [`comment`];
+
+// Значения
+const BAR_HEIGHT = 50;
+const SHOWING_FILMS_COUNT = 5;
+
+// Сервер
+const AUTHORIZATION = `Basic hrguy43grgh`;
+const LINK = `https://11.ecmascript.pages.academy/cinemaddict/`;
+
+export {
+  BODY_ELEMENT,
+  HIDDEN_CLASS,
+  ENTER_KEY,
+  ESC_KEY,
+  ACTIVE_CLASS_FILTER,
+  ACTIVE_CLASS_SORT,
+  BAR_HEIGHT,
+  AUTHORIZATION,
+  LINK,
+  COMMENT_FORM_FIELDS,
+  SHOWING_FILMS_COUNT,
+};
